Apply Handlebars
