+++ README.md
@@ -1,4 +1,6 @@
 # bulletin-board
 
 - Java: 1.8
-- Spring Boot : 2.1.5.RELEASE 
+- Spring Boot : 2.1.5.RELEASE
+- Gradle : 5.4.1
+- Handlebars: 0.3.0
