+++ build.gradle
@@ -15,8 +15,10 @@ repositories {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    api 'pl.allegro.tech.boot:handlebars-spring-boot-starter:0.3.0'
+
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    compile 'com.h2database:h2'
+    runtimeOnly 'com.h2database:h2'
     implementation 'org.springframework.boot:spring-boot-devtools'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
