+++ src/main/java/springbom/bulletinboard/business/ArticleBusinessService.java
@@ -1,7 +1,7 @@
 package springbom.bulletinboard.business;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 import springbom.bulletinboard.data.ArticleRepository;
 import springbom.bulletinboard.model.Article;
 
@@ -10,14 +10,18 @@ import java.util.List;
 @Component
 public class ArticleBusinessService {
 
-    @Autowired
-    ArticleRepository repository;
+    private ArticleRepository repository;
+
+    public ArticleBusinessService(ArticleRepository repository) {
+        this.repository = repository;
+    }
 
     public List<Article> retrieveAllArticles() {
         // 이곳에 비즈니스 로직을 추가할 수 있다
         return repository.findAll();
     }
 
+    @Transactional
     public Article addArticle(Article article) {
         return repository.saveAndFlush(article);
     }
